requirements_dev.txt
ipython
django-extensions
flake8
isort